feat: experimental tile highlight
It's hard to use mouseover event on svg when other objects are covering
targeted object. It's possible to change pointer-events on fly, but then
moving piece is another problem, so i moved "hover logic" to scala. All
changes are send via mousemove event on svg, and every time whole state
is being returned so its some performance hit. It's adding about 3-4ms
of computations to each mosemove event on my computer. As this feat is
not really needed I'm abandoning it.